backend/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
python-decouple
psycopg2-binary
gunicorn
whitenoise
pandas
openpyxl
